+++ eras/shelley/impl/CHANGELOG.md
@@ -32,6 +32,7 @@
   * `propWits` - will become an internal function in the future version
 * `validateNeededWitnesses` no longer accepts `witsVKeyNeeded` as an argument.
 * Move `ConstitutionalDelegCert` from `cardano-ledger-core` as `GenesisDelegCert`.
+* Fixed `NoThunks (ShelleyGenesis c)` instance, as it was incorrectly disallowing thunks in its `sgInitialFunds` and `sgStaking` fields
 
 ## 1.2.0.0